MLBackend/requirements.txt
numpy==1.21.6
pandas==1.3.5
scikit-learn==0.24.2
joblib==1.1.0

fastapi==0.70.0
uvicorn==0.15.0
pydantic==1.8.2
prometheus_client==0.16.0
prometheus-fastapi-instrumentator==5.9.1

torch==1.10.0
torchvision==0.11.1

opencv-python==4.5.3.56
pillow==8.3.2
scikit-image==0.18.3

python-multipart==0.0.5
validators==0.18.2
requests==2.26.0
matplotlib==3.4.3
